+++ include/depthai-shared/common/ImgTransformation.hpp
@@ -69,6 +69,8 @@ class ImgTransformations {
 
     void setRotation(float angle, dai::Point2f rotationPoint, int newWidth = 0, int newHeight = 0);
 
+    void setScale(float scaleX, float scaleY);
+
     // API that is meant for performance reasons - so matrices can be precomputed.
     void setTransformation(std::vector<std::vector<float>> matrix, std::vector<std::vector<float>> invMatrix, RawImgTransformation::Transformation transformationm, int newWidth, int newHeight);
 
@@ -82,72 +84,5 @@ class ImgTransformations {
 DEPTHAI_SERIALIZE_EXT(ImgTransformations,
                       transformations,
                       warpEnabled);
-
-std::vector<std::vector<float>> ImgTransformations::getFlipHorizontalMatrix(int width) {
-    auto scale = matrix::createScalingMatrix(-1, 1);
-    auto translate = matrix::createTranslationMatrix(width, 0);
-    return matrix::matMul(translate, scale);
-}
-
-std::vector<std::vector<float>> ImgTransformations::getFlipVerticalMatrix(int height) {
-    auto scale = matrix::createScalingMatrix(1, -1);
-    auto translate = matrix::createTranslationMatrix(0, height);
-    return matrix::matMul(translate, scale);
-}
-
-std::vector<std::vector<float>> ImgTransformations::getRotationMatrix(int px, int py, float theta) {
-    auto translateToOrigin = matrix::createTranslationMatrix(-px, -py);
-    auto rotate = matrix::createRotationMatrix(theta);
-    auto translateBack = matrix::createTranslationMatrix(px, py);
-
-    auto temp = matrix::matMul(translateToOrigin, rotate);
-    return matrix::matMul(temp, translateBack);
-}
-
-std::vector<std::vector<float>> ImgTransformations::getScaleMatrix(float scaleX, float scaleY) {
-    return matrix::createScalingMatrix(scaleX, scaleY);
-}
-
-void setPadding(float topPadding, float bottomPadding, float leftPadding, float rightPadding) {
-    return;
-}
-
-void setCrop(int topLeftCropX = 0, int topLeftCropY = 0, int bottomRightCropX = 0, int bottomRightCropY = 0) {
-    return;
-}
-
-void setFlipVertical() {
-    return;
-}
-
-void setFlipHorizontal() {
-    return;
-}
-
-void setInitTransformation(int width, int height) {
-    return;
-}
-
-void setRotation(float angle, dai::Point2f rotationPoint, int newWidth = 0, int newHeight = 0) {
-    return;
-}
-
-// API that is meant for performance reasons - so matrices can be precomputed.
-void setTransformation(std::vector<std::vector<float>> matrix, std::vector<std::vector<float>> invMatrix, RawImgTransformation::Transformation transformationm, int newWidth, int newHeight) {
-    return;
-}
-
-dai::Point2f clipPoint(dai::Point2f point, int imageWidth, int imageHeight, bool &isClipped) {
-    return point;
-}
-
-dai::Point2f transformPoint(RawImgTransformation transformation, dai::Point2f point) {
-    dai::Point2f returnPoint(0, 0);
-    return point;
-}
-
-dai::Point2f invTransformPoint(RawImgTransformation transformation, dai::Point2f point) {
-    return point;
-}
 
 }  // namespace dai
